Refactor config file
