+++ .gitignore
@@ -105,7 +105,7 @@ dist
 
 
 # Project data
-config.json
+config.ts
 data.db*
 *.service
 *.png
